feat: 添加 Java 代码判题功能，重构相关类，优化配置和接口

+++ zzcoder-judger/src/main/java/com/zazhi/judger/common/pojo/JudgeResult.java
@@ -18,8 +18,8 @@ public class JudgeResult {
     private List<TestCaseResult> testCaseResults; // 测试用例结果
     private Boolean fullJudge; // 是否全量评测
     private String errorMessage; // 错误信息
-    private Integer timeUsed = 0; // 时间
-    private Double memoryUsed = 0.0; // 内存
+    private Long timeUsed = 0L; // 时间
+    private Long memoryUsed = 0L; // 内存
 //    private String result;
 //    private int exitCode; // 容器退出代码
 }
